Start converting things to use parameters.

+++ src/params/parameter_repository.cpp
@@ -1,15 +1,15 @@
 #include "params/parameter_repository.hpp"
 
-void ParameterRepository::def(char const *name, const double value) {
+void ParameterRepository::def(char const *name, const float value) {
     if(parameters.count(name) == 0) {
         parameters[name] = value;
     }
 }
 
-void ParameterRepository::set(char const *name, const double value) {
+void ParameterRepository::set(char const *name, const float value) {
     parameters[name] = value;
 }
 
-double ParameterRepository::get(char const *name) const {
+float ParameterRepository::get(char const *name) const {
     return parameters.at(name);
 }
